Lots of wrapping unsupported stuff in WINVER #ifs.  This is to get a build working on windows XP.
Define WINVER as a preprocessor variable in the project files for this
to work.  Windows XP is 0x0501 I think.

Had some oddness with uint8_t on linux

One of the changes is to the shell's get folder path stuff.  That
changed sometime after windows XP, so I tried to translate from the new
folder ids to the old CSIDL dirs.